requirements.txt
pandas
plotly
python-dotenv
Requests
openpyxl
numpy

